requirements.txt
nltk==3.2.5
python-Levenshtein==0.12.0
six==1.11.0
